move menus up
